0c3ca8798c143828d64b2bbbb13bba7c44fc9313 refactor(api): [#143] remove duplicate code (Jose Celano)
1c72ac078bca8e5750d35f82184504ee891304d5 docs(api): [#143] remove deprecated comment (Jose Celano)
b7c514431d822dfa494fe4fd6925f5241d92de44 refactor(api): [#143] change fn return type (Jose Celano)
02dfe3ee87acaa20b4ed4b4432dab2ca9a5e8b34 refactor(api): [#143] rename response functions (Jose Celano)
6955666bf0d7be54d9bae8936237f7d9608ac8ba docs(api): [#143] remove comment (Jose Celano)
6ddbdd95c8fe2f96e306e753d5b3fa3bb1461838 docs(api): [#143] move comment (Jose Celano)
0940463449a314bd35eec17caf0d31f0ae9283a9 refactor(api): [#143] extract api mods (Jose Celano)
2a92b0a01a1f199dab23bfef3ffbd3ee0a9cb272 refactor(api): extract mod for API responses (Jose Celano)
6dd3c482868dbee638e71b4bee86cce5c751d476 refactor(api): [#143] move API resources mod (Jose Celano)
77ec52184d4ad2eaaf22dfc0838e1568013b40ee refactor(api): [#143] remove Warp API implementation (Jose Celano)
337e12eef5a205bb772dbda24b0f5bb4c42df45a feat(api): [#143] replace Warp API with Axum implementation (Jose Celano)
8d32628e6e80bd5429917ea12e6e815577b0f203 refactor(api): [#143] extract and rename functions (Jose Celano)
072f3d7d42350ae157817843779499d7bf09587a test(api): [#143] add more tests for invalid key id URL path param (Jose Celano)
2c222ee906e7e52dd01a41d75f105a67233172c9 test(api): [#143] add test for invalid key duration URL path param (Jose Celano)
39c15c6bcf4690f16617cb92a3fd1094a9e71192 test(api): [#143] add test for invalid infohash URL path param (Jose Celano)
aa2a2ef4346044aabaf2c8bf0f7966b2b3f94bf8 fix(api): [#143] do not fail trying to remove a whitelisted torrent twice (Jose Celano)
3bcbbc9a1c784276f6e6dad8dc3b27ca9f7adee7 refactor(api): [#143] normalize test names for errrors (Jose Celano)
2da0719966e4c8e42ce277698c9daad95a159423 test(api): [#143] add tests for authenticaation (Jose Celano)
517ffde147ebbc4040967803af61f387a7c85722 fix(api): [#143] fix new Axum API enpoint when URL params are invalid (Jose Celano)
c502c1d939ef9d51c0457c5b3ebb73d56a04d9b0 refactor(api): [#143] remove duplicate definition of axum router (Jose Celano)
504cb9e8d6ce24eed5e7e0e5dc24b673ced3541d feat(api): the new Axum api uses the URL prefix /api too. (Jose Celano)
5d9dd9d90164e59915a200d66b5f2a293cbd38d1 refactor(api): extract asserts in tests (Jose Celano)
03ba166bdc8ebed49305613a175525cab324aea3 feat(api): [#143] axum api. GET /api/keys/reload endpoint (Jose Celano)
6b2e3bcfb55b75de20af4505f0f45d680b64da25 feat(api): [#143] axum api. DELETE /api/key/:key endpoint (Jose Celano)
0282e33931718e71ab4f2b05b9948cc50ccc91b1 feat(api): [#143] axum api. POST /api/key/:seconds_valid endpoint (Jose Celano)
a58d8310fd36a99e60b05ce61ccff85f5632cc71 feat(api): [#143] axum api. GET /api/whitelist/reload endpoint (Jose Celano)
2ddf2684b439d28ba9fd4743e13607472e688ee1 feat(api): [#143] axum api. DELETE /api/whitelist/:info_hash endpoint (Jose Celano)
5c5fcbd26f7e5f03d02d80105b85eaf6eb0d4b4d feat(api): [#143] axum api. POST /api/whitelist/:info_hash endpoint (Jose Celano)
e1ed929f1cf0abf450e246413536530accc4d904 test(api): [#143] add tests for database failure (Jose Celano)
1515753b3f2365aad5d646e3efeced1b608b15e9 feat(api): [#143] axum api. GET /api/torrents endpoint (Jose Celano)
c36b121dce4674d43011920c56189facf5972bc5 refactor(api): [#143] extract service tracker::services::torrent::get_torrents (Jose Celano)
a8061792f4ae1b9fcb95f778d42f706f05206ab5 refactor(api): [#143] use extracted service in the Warp handler (Jose Celano)
ded4d110dc04b71d6448a6dc3c51a6049acf37e3 fix: clippy errors (Jose Celano)
a649fe861a5f7508621c4c7630c676dc28c68d8d feat(api): [#143] axum api. GET /api/torrent/:info_hash endpoint. Not found case (Jose Celano)
2aebf9ad396298d41454bfda75528395fab0d085 test(api): [#143] add test for torrent not known response in GET /api/torrent/:info_hash endpoint (Jose Celano)
16d438dced06f481744160c9199e20fd893822ab feat(api): [#143] axum api, WIP. GET /api/torrent/:info_hash endpoint (Jose Celano)
fe4303c0af047add8b413e8cb8b4f980e58b7e8c feat(api): [#143] SSL support for the new Axum API (Jose Celano)
af51f77743e58c3f0fb9121a97447e90a7b9ba09 feat(api): [#143] add new cargo dependency: axum-server (Jose Celano)
13959452f59949b91789050b8eda597fb82728b7 refactor(api): [#143] remove dummy api endpoint (Jose Celano)
43dbed933c5a7560fc36aa2fde981d89020d39bc feat(api): [#143] authentication with GET param for Axum API (Jose Celano)
1c6db6e6c47c8ed5ebd5efba8c1a0a541eb93a3f test: [#143] add tests for extracted functions (Jose Celano)
0f99f7bc077d6b15494a910c930e4045f3db5613 refactor: [#143] remove duplicate or unneeded code (Jose Celano)
0615c9f028853285fbc6aabac604a623c4d278bc refactor(api): [#143] remove duplicate code (Jose Celano)
7331c82d13f7ef01028c5600ecaa966ac73ef955 refactor: [#143] replace unwrap with expect (Jose Celano)
6a9e2d5a9f1396e4486b95d2e187740f368549bf feat(api): [#143] axum api, GET /stats endpoint (Jose Celano)
5ee3f93fb515cc26a43a5cab45d84a2011c0cbb7 refactor(api): [#143] extract mods for API testing (Jose Celano)
cbf88377950d0208cf3f4d8641978cbd1a1a823f feat(api): [#143] scaffolding for new API using Axum (Jose Celano)
901bc342721d137c9ca2b571a7e5632c33142b4b feat: [#143] add axum dependency (Jose Celano)

Pull request description:

  **UPDATED**: 12/01/2022

  The tracker API uses the web framework [Warp](https://github.com/seanmonstar/warp).
  This PR replaces Warp with [Axum](https://github.com/tokio-rs/axum).

  ### Tasks

  - [x] Basic Axum configuration.
  - [x] Authentication using token in GET params.
  - [x] Enable SSL.
  - [x] Add tests for the current API. Error cases are not tested. I think all cases are when there is a database error, and it returns an `ActionStatus::Err` error with a message.
  - [x] Reimplement endpoints. See subtasks.
  - [x] Improve asserts by checking the response body for OK (`200`) responses.
  - [x] Use the same URL prefix as the current API `/api/xxx`.
  - [x] Remove duplicate definitions of routes in Axum server (`start` and `start_tls`)
  - [x] Add tests for the current API. Error cases when GET params cannot be parsed are not tested. `info_hash` and `seconds_valid` Path params, and `offset` and `limit` GET params.
  - [x] Fix a bug with the auth `token`. The token must be the first param in the `query` otherwise, it does not work. In the end, it was not a bug. It only happens with `curl`.
  - [x] Add test for the current API. When you try to remove a torrent from the whitelist twice, you get a 500 response with this body
  : `Unhandled rejection: Err { reason: "failed to remove torrent from whitelist" }`
  - [x] Add some tests for malicious user input.
  - [x] Remove the Warp implementation.

  ### Endpoints subtasks

  Stats:

  - [x] `GET /api/stats`

  Torrents:

  - [x] `GET /api/torrents?offset=:u32&limit=:u32`
  - [x] `GET /api/torrent/:info_hash`

  Whitelisted torrents:

  - [x] `POST   /api/whitelist/:info_hash`
  - [x] `DELETE /api/whitelist/:info_hash`

  Whitelist commands:

  - [x] `GET /api/whitelist/reload`

  Keys:

  - [x] `POST   /api/key/:seconds_valid`
  - [x] `DELETE /api/key/:key`

  Key commands
  - [x] `GET /api/keys/reload`

ACKs for top commit:
  josecelano:
    ACK 0c3ca8798c143828d64b2bbbb13bba7c44fc931

Tree-SHA512: 592930e7e0739de0038e2511e04668b42bba632b8531e60298e5092e68a8dc252a12604f3fa93157595cb0f0ab71a12b06fdfa324eeaa5762b45232a11523aac

+++ src/tracker/mod.rs
@@ -1,6 +1,7 @@
 pub mod auth;
 pub mod mode;
 pub mod peer;
+pub mod services;
 pub mod statistics;
 pub mod torrent;
 
@@ -25,7 +26,15 @@ pub struct Tracker {
     torrents: RwLock<std::collections::BTreeMap<InfoHash, torrent::Entry>>,
     stats_event_sender: Option<Box<dyn statistics::EventSender>>,
     stats_repository: statistics::Repo,
-    database: Box<dyn Database>,
+    pub database: Box<dyn Database>,
+}
+
+#[derive(Debug, PartialEq, Default)]
+pub struct TorrentsMetrics {
+    pub seeders: u64,
+    pub completed: u64,
+    pub leechers: u64,
+    pub torrents: u64,
 }
 
 impl Tracker {
@@ -86,6 +95,7 @@ impl Tracker {
     ///
     /// Will return a `key::Error` if unable to get any `auth_key`.
     pub async fn verify_auth_key(&self, auth_key: &auth::Key) -> Result<(), auth::Error> {
+        // todo: use auth::KeyId for the function argument `auth_key`
         match self.keys.read().await.get(&auth_key.key) {
             None => Err(auth::Error::KeyInvalid),
             Some(key) => auth::verify(key),
@@ -121,7 +131,9 @@ impl Tracker {
 
     /// It adds a torrent to the whitelist if it has not been whitelisted previously
     async fn add_torrent_to_database_whitelist(&self, info_hash: &InfoHash) -> Result<(), databases::error::Error> {
-        if self.database.is_info_hash_whitelisted(info_hash).await.unwrap() {
+        let is_whitelisted = self.database.is_info_hash_whitelisted(info_hash).await?;
+
+        if is_whitelisted {
             return Ok(());
         }
 
@@ -140,9 +152,28 @@ impl Tracker {
     ///
     /// Will return a `database::Error` if unable to remove the `info_hash` from the whitelist database.
     pub async fn remove_torrent_from_whitelist(&self, info_hash: &InfoHash) -> Result<(), databases::error::Error> {
-        self.database.remove_info_hash_from_whitelist(*info_hash).await?;
-        self.whitelist.write().await.remove(info_hash);
+        self.remove_torrent_from_database_whitelist(info_hash).await?;
+        self.remove_torrent_from_memory_whitelist(info_hash).await;
         Ok(())
+    }
+
+    /// # Errors
+    ///
+    /// Will return a `database::Error` if unable to remove the `info_hash` from the whitelist database.
+    pub async fn remove_torrent_from_database_whitelist(&self, info_hash: &InfoHash) -> Result<(), databases::error::Error> {
+        let is_whitelisted = self.database.is_info_hash_whitelisted(info_hash).await?;
+
+        if !is_whitelisted {
+            return Ok(());
+        }
+
+        self.database.remove_info_hash_from_whitelist(*info_hash).await?;
+
+        Ok(())
+    }
+
+    pub async fn remove_torrent_from_memory_whitelist(&self, info_hash: &InfoHash) -> bool {
+        self.whitelist.write().await.remove(info_hash)
     }
 
     pub async fn is_info_hash_whitelisted(&self, info_hash: &InfoHash) -> bool {
@@ -277,6 +308,27 @@ impl Tracker {
         self.torrents.read().await
     }
 
+    pub async fn get_torrents_metrics(&self) -> TorrentsMetrics {
+        let mut torrents_metrics = TorrentsMetrics {
+            seeders: 0,
+            completed: 0,
+            leechers: 0,
+            torrents: 0,
+        };
+
+        let db = self.get_torrents().await;
+
+        db.values().for_each(|torrent_entry| {
+            let (seeders, completed, leechers) = torrent_entry.get_stats();
+            torrents_metrics.seeders += u64::from(seeders);
+            torrents_metrics.completed += u64::from(completed);
+            torrents_metrics.leechers += u64::from(leechers);
+            torrents_metrics.torrents += 1;
+        });
+
+        torrents_metrics
+    }
+
     pub async fn get_stats(&self) -> RwLockReadGuard<'_, statistics::Metrics> {
         self.stats_repository.get_stats().await
     }
@@ -308,5 +360,53 @@ impl Tracker {
                 torrent_entry.remove_inactive_peers(self.config.max_peer_timeout);
             }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use super::statistics::Keeper;
+    use super::{TorrentsMetrics, Tracker};
+    use crate::config::{ephemeral_configuration, Configuration};
+
+    pub fn tracker_configuration() -> Arc<Configuration> {
+        Arc::new(ephemeral_configuration())
+    }
+
+    pub fn tracker_factory() -> Tracker {
+        // code-review: the tracker initialization is duplicated in many places. Consider make this function public.
+
+        // Configuration
+        let configuration = tracker_configuration();
+
+        // Initialize stats tracker
+        let (stats_event_sender, stats_repository) = Keeper::new_active_instance();
+
+        // Initialize Torrust tracker
+        match Tracker::new(&configuration, Some(stats_event_sender), stats_repository) {
+            Ok(tracker) => tracker,
+            Err(error) => {
+                panic!("{}", error)
+            }
+        }
+    }
+
+    #[tokio::test]
+    async fn the_tracker_should_collect_torrent_metrics() {
+        let tracker = tracker_factory();
+
+        let torrents_metrics = tracker.get_torrents_metrics().await;
+
+        assert_eq!(
+            torrents_metrics,
+            TorrentsMetrics {
+                seeders: 0,
+                completed: 0,
+                leechers: 0,
+                torrents: 0
+            }
+        );
     }
 }
